Updated recipe for (python3-)inotify to use pypi
For the sake of consistency with development, inotify is not provided using a pypi-based
recipe, which has also been a good opportunity to learn how to construct a simple one.

This results in a similar file to git-based versions, except PV is used for versions. Also,
the contents of the src .tar.gz can differ from the original git source. In this example,
the LICENSE file does not appear to be in the pypi-distributed source, and so BitBake
requires the license to be set as CLOSED.