Configured U-boot using devtool flow to reset first boot device
The devtool flow uses a workspace directory to change project components.

When these changes are made, the results are then automatically added as patches to the meta-user recipe layer,
which is tracked in this repo. Care should be taken to make sure this step is performed, since it can easily
be missed (the workspace is automatically added as a bitbake layer, so for the local build will work, but changes
will not be seen on the remote repository).