Corrected issue where development host IP was being built into project
When the u-boot configuration switches from other to the PetaLinux variant, it seems to automatically include a definition
for assigning the u-boot TFTP server ID, with the default being AUTO, a value which causes the IP to be taken directly
from the development host, which then appears in platform-auto.h in the user layer.

Changing the value in the config to empty clears the IP on re-generation of the platform-auto.h file.