Updated Vivado block design and constraints to match Tim's 2018 Zynq mods
Modifications have been reproduced from the 2018 toolchain version to add connections to the block design
for I2C and SPI interfaces. The process appears to be exactly the same.

The changes have also been exported to the block_design .tcl script using
in Vivado, which results in the changes being re-created when the Trenz scripts are used to build the
whole project from scratch.