Update configuration to include options specified as PetaLinux modifications from Trenz
As mentioned before, Trenz specified some changes that should be included in PetaLinux projects.
If the default PL project is used they are already present, but must be added manually after the creation of a new
project, which was missed when I generated a new one for the 2020 toolchain.

Most of the changes to config are related to I2C and EEPROM access, including for the MAC address.