Re-include HW modifications to add SPI and I2C
Since the modifications to the hardware had not been removed in the Trenz source,
including them in the PetaLinux build was as easy as re-generating the files in
Vivado and Vitis and re-exporting and importing them to the PetaLinux Project.

Modification was also made to the FSBL source to prevent two 0x50 EEPROMs being
MUXed in at the same time. Note that these changes are made in the /sw_lib/
directory, as the /workspace/ directory is cleared and regenerated from the SW
library folder on project (re-)creation.