Update autobuild.sh to abort duplicate XSA import
A lof of confusion was caused in the project when files in project-spec/
hw-description/ are apparently erased if an XSA the same as the current
one is imported. The script now checks if these are the same, and aborts
if they are (even if the name is different).

Also, to prevent the current XSA being deleted from the PetaLinux cwd
without replacing it, the delete is only performed after the Trenz XSA
is found and checked.