Update HW design to slow SPI base clock to 1Mhz, add dtc
The SPI base clock is now 1MHz to make development easier. The SPI driver max
clock divider is 256, therefore slowest clock is 3.9kHz. This can easily be
reversed later on if a faster clock is required.

dtc has also been added to the image recipe, since it is useful for printing
out the in-use device-tree passed to the kernel during runtime (and in QEMU).