Update odin-sequencer recipe to match latest build
odin-sequencer had been updated, but since the recipe was set to build
the latest commit on git, it immediately pulled changes. Unfortunately,
the setup.py patch was then no longer compatible, and a build failure
was caused.

The setup-py patch for odin-sequencer has now been updated to reflect
the changes, and instead of building latest, the recipe is now set to
build at odin-sequencer commit 6128e0c. This can be updated manually
for future versions, as long as compatibility is checked first.