fixed contraints and added spi device tree changes
Unfortunately the comments added to the constraints file fail to meet the
allowed syntax, so have been removed. The new PMUFW, FSBL were built an
included in the build along with the newly compiled .xsa that was extracted
to the project-spec/hw-description/ folder.

The device tree has also been updated to include the new chip select device
on SPI0 and the new bus for the MERCURY ASIC spi1