Update HW and image to support 12 EMIO GPIO lines
Still needs further work, but EMIO has now been added to the Linux image
with a width of 12. The input and output works though chardev (new) and sysfs (old).

libgpiod was added, and has been forced to compile the python (3) bindings so that
gpiod can be imported in python and used to control the GPIO lines exposed.

Note that offsets can make things confusing. When using sysfs, there is an offset of
338 because of the driver base. There is also an offset of the MIO lines to get to
the EMIO of 78. When using chardev (gpiod) the base is ignored, but the MIO offset
remains. Therefore EMIO pin 0 is at 416 for sysfs and 78 for chardev (gpiod).

A gpio driver for odin-devices is currently in development by me to abstract away the
offset for different platforms, like Zynq (which has MIO width of 54 as opposed to 78
for ZynqMP).