libgpiod version updated from 1.4.1 to 1.6.2
The libgpiod version 1.4.1 was too old to included the python bindings for the Line.update()
function, which is required for syncing the line information with the kernel. The update was
simply a matter of adding a recipe to override the 1.4.1 version and adding the append from
the previous commit to make sure the python bindings are compiled in.

This change has been tested with the gpiod python3 bindings and has been verified to correct
sync issues between jobs in QEMU. It has not been tested on hardware.