Add LVDS (1v8) pair as GPIO0
An LVDS pair has been added to the GPIO interface, and can be written
to (output only) on GPIO0. As far as the kernel is concerned, this is
one output line since it has been buffered from EMIO in Vivado.

This LVDS pair now appears on AD7 (P) and AE7 (N), and under test has
shown an output Vod of 370mV and Vcm of 1.215V with a 100ohm
termination. This is within the expected ranges documented in the AC
DC Switching Datasheet for UltraScale+ MPSoC.