Changed recipes to use specific version filenames
Unfortuantely, wildcard filenames for versions of software is supported only by the
.bbappend files. This is so that they can apply to multiple recipe files. The recipe
must refer to a specific version, or a git commit hash with AUTOINC (may require more
investigation). For now, there will be specific recipes for each version.

Due to the fact that this will result in more files per recipe as the versions are
added, the components of ODIN have been separated into their own recipe directories.