Add new odin-control management service
This service allows the odin-control detector instance (not necessarily HEXITEC-MHz) to be started,
stopped and restarted with ease. The interface for this is generic so that other detectors can just
require this additional recipe. Since it has superseeded the original startup script, this has been
removed.

Additionally, this setup system allows a configuration file to be placed on the mmc (if desired) so
that the default installation can be modified or bypassed completely for easier development. Currently,
if the directory /mnt/sd-mmcblk1p1/loki-config is found, the default config will be placed inside
automatically for free modification. Additionally, a setup script can be added for further
customisation. For example, the script could mount network storage and indicate that a virtual
environment with different versions of odin-control, odin-devices etc should be used, with other
static resources in a development directory.

The service restart can be used to accept new configuration on the fly without reboot. The configuration
is non-volatile since it is stored on the MMC.

Further documentation on this will follow.