Add two mechanisms for network configuration from external SD
The simplest method is now to directly set a static IP, which will be performed
before the init script is run, meaning any network mounts that are made within
the aforementioned script will be performed with the new interface configuration.

For more complex configurations, the network config file usually found at /etc/
network/interfaces can now be replaced by a file at <loki-config>/interfaces, if
the switch in the main configuration file is active. This interfaces file should
be created by the image to match the internal file if it did not yet exist at
system boot time.