Auto-mount flash mtd partition and use for dropbear ssh key
A new recipe has been added to add an entry into /etc/fstab to mount the 'bootenv'
mtd partition (otherwise unused) as a JFFS2 filesystem. The mountpoint is also
created in the image.

Dropbear now points its key location (using the recently added bbappend) to this
non-volatile filesystem, meaning that the key generated on first boot of the device
is persistent.