Limited files owned by loki, moved chown
The ownership chown commands are now performed from the loki recipe, since this
should not be in the control of the detector recipe (yet). If it is later needed,
the detector recipe could own directories by including USERADD_PARAM and inheriting
from useradd.

The number of files owned by loki has also been reduced. It is not a requirement
that the detector odin instance files are owned for them to be read and used for the
server. However, a new space at /opt/loki-detector/exports has been reserved in case
the server needs to create files somewhere (temporarily, since it is volatile).