Added Makefile
Added Makefile to replace existing autobuild script, which operates controls the
PetaLinux toolchain from importing the Vivado/Vitis exported files to building,
and packaging. This is far more flexible than the existing system since it will
only re-run operations when they are required due to changes. The makefile will
asssume that the hardware, FSBL and PMUFW have already been generated for now.

The build depends on both the user layer and the yocto workspace, so that if a
recipe has been exported for customisation it will still trigger a build.