Support Vivado Project creation and deletion in Makefile
The top-level makefile is now capable of generating the Vivado project through
the Trenz scripts from a fresh repository. Additionally, a recipe for exporting
changes in the Vivado project out to the block design TCL script so that they
can be added to version control. While this is not a requirement of 'make all',
running 'make clean' will remove the built binaries, clean the OS, and delete
the Vivado project (after checking for changes and automatically exporting them
to the block design TCL). 'make mostlyclean' will simply delete the generated
HW/SW and clean the OS.