Add ssh authorized_keys persistence in mtd flash
By default, the .ssh directory for the loki user is now mounted from the flash partition, meaning
that keys added to authorized_keys will remain persistent through both device reboots and image
updates. Since it is possible that default keys may be added to the image in the future, any keys
present in the image's default .ssh directory for loki will be copied into the flash partition's
copy automatically.

Additionally, there is a new configuration flag in the external config file that allows this
behaviour to be disabled, relying only on the image's .ssh directory. To stop ssh persistence,
simply set conf_PERSISTENT_SSH_AUTH to 0.