V0.1.3 staging (#39)
* Add ssh authorized_keys persistence in mtd flash (#37)

By default, the .ssh directory for the loki user is now mounted from the flash partition, meaning
that keys added to authorized_keys will remain persistent through both device reboots and image
updates. Since it is possible that default keys may be added to the image in the future, any keys
present in the image's default .ssh directory for loki will be copied into the flash partition's
copy automatically.

Additionally, there is a new configuration flag in the external config file that allows this
behaviour to be disabled, relying only on the image's .ssh directory. To stop ssh persistence,
simply set conf_PERSISTENT_SSH_AUTH to 0.

* Fix/makefile tweaks (#38)

* Add rootfs dependency for petalinux-package

* hardware dependency fix and DTS name fix

Fixed dependency for hardware so that design changes require a build and moved DTS/DTB definition so that it is actually included as part of the all build

* Redirect odin_server stderr

Previously, issues that crashed odin_server were hard to diagnose from the logs because they didn not contain the errors that caused the crashes. Now, stderr is recorded to a configurable output file. By default, this is the same file as the rest of the output.

* Move the logging to a directory in /var/log, so that new files can be created by the loki user without permission issues. This allows the log rotation to function.

* Add numpy to build for general use

* Update mercury-detector commit to include latest UI edits

* Temporary change to odin-devices target to use add-ltc2986 branch

This branch is now a requirement for ASIC diode monitoring

* Minor fix to makefile

* Improve makefile Vivado design source checking

Previous design directory checking had false positives that resulted in builds
being forced without being necessary. Now, the dependency is on specific source
file locations and searches for files only (since Vivado seems to have a habit
of updating directory timestamps without the files inside the having changed).

* Add matplotlib support

* Update mercury-detector to version with plotted segments

Additionally, add the images output directory in the web server's static area,
so that they can be easily loaded by the browser. This has to be writeable by
the restricted loki user, while the rest of the static area is not.

* Bump mercury-detector for vmaxmin fields

* Bump HEXITEC-MHz ODIN to include cal tests and safe start