Added beginnings of alternative carrier support
There will be base support block designs and constraints structures used for different carriers
to provide a working image. This will later be supplemented later on (in the application starter
kit) with additional application constraints files, and the option to override the default block
design and constraints directory if desired, though ideally this should not be required for most
basic applications. This will require updates to the block design to allow it to provide more
generic functionality.