Added folder for carrier-specific prebuilt hardware
Eventually, although this should not be included in the main repository tree, the releases
should include a version that can be downloaded with prebuilt hardware for all carriers and
all carriers. There should be a rule in the loki makefile to produce all of these. This would
be the basis of an application project, especially one that did not (at least initially) require
custom hardware.

This would also mean that the user should not require anything other than the petalinux tools
installed to complete an application build.