Update LVDS and GPIO generic lines, block design fixes
The LVDS and generic GPIO lines have now been added to constraints. Although
these have no application in the base design (and are potentially application
specific), they are provided so that there will be some GPIO capability in a
base image for debugging and test, without the application needing to make any
modifications to the block design in the first instance.

Having fixed many bugs with the constraints and block design, this is the first
of the neo-loki commits that successfully builds the hardware.