Update autoconf to use config.site for variable defaults
Previously, variable values set with AC_SUBST had been assumed to be defaults, overridden
with the call to ./configure variable=x. However this was not the case; the assignments
in the configure.ac always took precendence.

Still wanting to have defaults hard-coded for the project for certain variables, I have used
a config.site file to check the variables during the configuration, and set defaults if they
are not otherwise set. This also allows some checking to be performed, such as throwing errors
if 'required' variables are not set.