Added additional makefile for autoconf
This file is designed only to provide the means to conveniently configure the LOKI project from a calling application makefile without requiring specifics of the autoconf being needed in the application makefile. The application can now include this file, and call loki-configure-hw/sw/os depending on the recipe in use, and can specify the configuration arguments required in a variable. See HEXITEC-MHz example.