Added rules, fixed prefix issue with making Makfiles with autoconf
Autoconf was failing to make some Makefiles if they did not exist, which turned out to be
due to my attempt to use several rules for the same recipe in a manner that did not make
sense, but didn't cause an error. This has now been corrected.

During this process the prefix handling has been improved a bit in the configuration
makefile, config.mk.

A few rules for cleanup have been tweaked.