Overhaul IO handling with PinHandler interface class
The interface for IO is now provided by a separate class that could be easier to swap out
in the future if required (for example for libgpiod2). Basic IO configuration management is
now performed with defaults in the code that can be overridden in the main odin config file.

These defaults can also be set in the derived extension classes for additional pins.

In addition, helper extension for LED and Button type pins have been added with more specified
defaults to that the carriers can define them quickly. Carrier 1v0 has been updated with these
changes.

The pin reading is currently direct for input pins, but should be updated to use a background
loop.

Pins for the application that require additional functionality like direct reading or events
etc should be handled separately with gpiod. This module's gpiod usage does not impact otherwise
not requested pins