Add build information including tag to etc and expose through loki adapter
It has previously been difficult to determine what version of software is running
on a given system. Now, some information about it has been build into the image
into /etc/loki/. 'application-name' and 'application-version' are supplied as
environment variables by the application repository.

'platform' is currently the combined carrier and module shortname. 'version' is the
git description of the current loki repo version when built. It is anticipated that
the application version will be implemented in a similar way, meaning that someone
with access to these values can rebuild or inspect exactly the same image.

In addition, the loki adapter pulls the values of these on instantiation, and exposes
them through the parameter tree.