Change Ethernet PHY Address to 0x00
The LOKI carrier sets the 88E1512 Ethernet PHY CONFIG pin low, meaning
that the PHY address will be 0x00 rather than 0x01. This therefore
differs from the Trenz TEBF0808 carrier's setting, which is not
configurable. The Ethernet PHY is now detected and works properly.

The issue with this is that this build will now not have functioning
ethernet for TEBF0808. Some method of altering the device tree
differently depending on the carrier selected will be needed at some
point to allow for changes like this without removing support for the
old platform.