Fix Micro SD (patched for broken CD and WP)
MicroSD has now been fixed in the Vivado design such that it can be
booted from and mounted from inside the Linux kernel.

Primary changes are:
- WP is not required (or even present) on uSD cards, however, if removed
from the Vivado design will always report as read-only. Therefore, it is
included on MIO44 (however this is not connected on the board) and
pulled down internally to indicate writes are allowed.
- Speed of the clock has been slowed to 100MHz as advised by Trenz
appnote.
- CD card detect has been disabled by both removing it in Vivado (so
that the FSBL does not require it) and in the device tree for the kernel
with 'broken-cd'.

The change to the CD line is present because of an issue on the PCB
design of the first revision where the CD is inverted (connected to 1v8
when a card is inserted). Although this can be corrected in the device
tree with 'cd-inverted', the device does not work in U-Boot, so CD has
been disabled. This therefore works well using polling (mounts still
appear on insert as far as the user is aware), however for the next
revision correcting it to read the line should be considered.