Add loop to service stop in init script to fix restart
The init script for LOKI already had a stop command to terminate the
currently running ODIN process. However, it had no form of blocking to
wait until the service had actually stopped before moving on. Previously
this happened so quickly it was not an issue, but now that the SIGTERM
is handled properly by odin-control 1.5.0, it performs cleanup
operations which can take time.

This essentially mean that the service restart command wasn't always
working because it would try and start the process before the cleanup
had finished, and then crashed because it was unable to get access to IO
devices.

There is now a blocking wait after the service stop. If the user is
impatient, they can CTRL-C and move on, but otherwise this fixes the
restart service command.