Added generic LTC2986 temperature monitor setup
The LTC2986 had previously not been configured at all, but is now
initialised as a device. Some configuration is optionally provided to
activate a PT100 temperature sensor (RTD) that can be mounted to the
board.

Otherwise, the setting up of sensor channels must still be done by the
user since it is application-specific and the channels are routed
off-board. For this, the device handler is provided via an accessor
function. There is also a function to read numbered channel values,
though it is direct and not rate limited to allow for maximum
flexibility. This means rate limiting must be observed in the
application code.