Add thread watchdog and simpler add_thread function
The process of adding a thread has been made easier, mostly for use by
derived classes by using an add_thread function to specify thread name,
target function and optional arguments.

There is also now a watchdog thread and associated functions that can be
used to (optionally) add a watchdog period to any thread added using the
above function. On failure by default this will report via the loop
status (parameter tree) and also report in the logs as an error.
However, there is optional support to register a callback for failure of
a given thread.