Update PL config to user remote kernel source 5.15
PetaLinux configuration has been updated to use a remote kernel source
rather than the toolchain-packaged 5.4 linux-xlnx version. It has now
been specified as version 5.15, still from the linux-xlnx repository
(the version included in the Xilinx 2022.1 release).

This is not a risk-free process, as the Yocto layers have been designed
without kernel upgrades in mind. The image does not yet build.