Override do_share_workdir in kernel Yocto Recipe
The Yocto meta layer includes a kernel class: kernel.bbclass, which is
inherited by linux-xlnx. In kernel versions 5.8+, the file
Module.symvers is not necessarily built. This requires a single line
change to the function (which was implemented in later Yocto layer
versions) that checks for the file presence before copying it, making it
backwards compatible.

Unfortunately through PetaLinux the class cannot be overridden directly,
and therefore the function has to be overridden in the kernel bbappend
file in the user layer. This is messy, but necessary. Later versions of
Yocto have this fix in the meta layer in the toolchain. This patch is
also backwards compatible.

In addition, no longer required kernel patch for SPI has been removed,
as it was fixed in this later kernel from Xilinx.