Add init script to auto-format empty eMMC
The eMMC is used as a boot and storage device. However, when new LOKI
carriers are received, they will be unformatted. This script detects
unformatted eMMC, formats it and then reboots.

This script is required by the main LOKI config script, so will always
run first, meaning that unless the eMMC device is simply broken, there
will always be eMMC present for the LOKI detector software.

This has also necessitated the inclusion of 'parted' in the image, which
could come in fairly handy anyway. Besides, it has little impact on
image size.