Add control host init script, useful aliases, system id init scripts
An additional init script for getting the system ID has been added. This
recovers the system ID (which is unique to a LOKI board / detector
system) from the eMMC module by default, therefore meaning they are
always unique without any interaction or bring-up. If desired, the ID
can be overridden by writing the config file loki-config/system-id in
the eMMC, which is more easily accessed with the loki_set_system_id
helper created (see below). This allows the user to set more
human-readable names, like 'mhz-system-1' for example.

Another init script for initiating connections with a control host has
been added. The IP and port of the host are configurable by setting
enviironment variables in the eMMC in loki-config/loki-host.conf. This
is a complex script that checks for presence of the host, and if found,
will find / create the application directory structure in the associated
network mount, and then scan for a layout.conf file. If not found, it
will dump an example file in the location and abort. If a layout file is
found, the directory will be traversed, and various directories, if
found, will be bind mounted onto the system.

This includes:
    - A data export directory, useful for data read over SPI, or any
    other file write from Python or otherwise (mapped to
    /opt/loki-detector/exports)
    - The sequences directory for odin-sequencer
    - An image update directory, for staging firmware updates. Currently
    there is no process for actually updating from here.
    - A general LOKI configuration file (loki-config.conf) override.

By default, these directories will be shared by any other detector
system of the same type (application) that is plugged in. However, if
enabled, the user can specify directories using ${LOKI_SYSTEMID}, which
will result in each system using its own ID in the path name, meaning
different systems could have different data output directories,
sequences etc.

Finally, loki-aliases.sh is a file that includes several useful aliases
and commands that make using the LOKI system easier over ssh. Currently
this must be manually sourced from /opt/loki-commands/loki-aliases.sh,
but I intend to include it automatically. It will be added to.