Upgrade network config for DHCP default with additional config in eMMC
The default interfaces file for network configuration is now configured
to set up only eth0 with dhcp by default.

Optionally, if there is ANY config file located in the eMMC in
/mnt/sd-mmcblk0p1/interfaces-mmc/*, it will be sourced by the network
init script. This means that the networking can be customised
completely.

It is suggested that if you want a fallback address, create an aliased
network with a new static address, which will allow for DHCP where
possible, and a fallback static address if connected to a machine
without DHCP.

To aid in creation of this config, a new helper command
loki_set_static_ip has been created, which will write a config file in
the eMMC to create a fallback address. To remove it, delete the file.