Re-write eMMC bootstrapping script and fix non-volatile net interfaces
It was found that the network interfaces directory automatically bind
mounted from the eMMC if present was not working, since although this
was added to the fstab, the fstab was processed before the udev rules to
mount the eMMC. Although it would have been possible to modify the
system init scripts, it was neater to change approach and upgrade the
existing eMMMC formatting init script to a more general 'bootstrapping'
script, which creates the bind mount before the network service runs.

This is still as flexible as before; if there is a directory, any conf
files inside it are sourced. If there is no directory, it is ignored
without issue (and the system will default to DHCP).

To accommodate this change, the eMMC bootstrapping script has also been
set to start on run level S, which ensures it is before the networking
script (without having to modify the system scripts at all).

Since it is now performed manually, the mount has been removed from the
fstab.