Added ecdsa host key generation and moved keys to eMMC
The previous system generated only RSA sha1 host keys on boot, which are
now likely to be automatically rejected by a host system. Therefore, the
image will now generate an RSA and ECDSA key on boot, and now store them
in the eMMC rather than the flash (since the SoM is now considered an
interchangeable part).

To accomplish this, it was necessary to alter the init script for
dropbear by dropping in a replacement script. The duplicates operations
undertaken for the RSA key for the ECDSA key, and registers both with
the daemon. In addition, keys are generated in /tmp/ and copied to the
eMMC because generating them directly to the eMMC fails.

This means that host systems are at liberty to use either newer ECDSA or
RSA at will.

The .ssh directory previously created on the flash and bind mounted to
the image for persistence of authorized_keys has also been moved to the
eMMC, but the mechanism is still the same. This takes place in
loki-config.conf.