Register Controller released on exception in context
Previously, when used with 'with', if an exception was encountered while
inside the managed context, the yield would return and raise it without
ever releasing the mutex. This meant that any error while reading the
ASIC fields / registers resulted in the whole system locking up.

For example, this was seen in the babyd-embedded implementation both
when disabling the ASIC interface during reset, and additionally when
attempting to write to any field that did not exist.

The errors are still raised, but the mutex is also handled properly with
a 'finally' block such that the system can continue running. The user
should still handle whatever exception has been raised.