Improve pin handler to allow for multiple gpiochips
Previously it had been assumed that there would only every be one
gpiochip. However, when multiple GPIO buses are added, each gets its own
chip. Therefore, the functions for adding pins have now been updated to
support having different specified chips.

- There is a 'base' chip used as the default chip number if one is not
given for each pin. The number used as the default can be overridden.
- Pin IDs supplied to 'add_pin' are now added in two ways, depending on
if the value has been determined to be numerical:
    - If numerical: the ID will be converted to it, and used with
    get_line() to recover the desired line. This REQUIRES a chip number,
    so either the default will be used, or the supplied chipnum passed
    to the function.
    - If a string: will be used with find_line(), which will find the
    desired name no matter which chip it may reside on.

This is backwards compatible with other adapters, as long as there is
only one gpiochip0. More specifically, this will not affect the
operation of the current HEXITEC-MHz or BabyD adapters.