Removed requirement to have all LOKI directories
The LOKI class previously assumed that one must have a static directory,
clock generator directory, config file for odin etc. However -especially
for initial tests- it is feasible that you would want to build an image
without all/any of these present.

Therefore, if the variables for their locations in your repo/files are
not defined, they will simply not be included.