Added recovery boot controlled by button
If button 1 is held down while U-boot starts running the boot.scr scipt,
the system will now attempt to boot an image from the QSPI flash, which
is now where a 'recovery' image will be kept. Otherwise the system will
boot from eMMC (assuming this is selected with the physical boot
switch).

In addition, U-boot now enables the board LEDs on boot, and uses them to
indicate when the script has started running (with a strobe). If the
recovery mode has been entered, it will leave two green LEDs lit (at
this point the user can stop holding the button).