Properly free mutexes in context manager on error
Similarly to a previous issue with register controller, if an error
occurred while a mutex was grabbed while using the context manager for a
device handler, it caused the system to never release the lock, which
has resulted in unpredictable lockups.

Therefore, if there is an error it is raised so that it can be handled,
but the mutex is also freed.