Add mechanism for image information from runtime and storage devices
There is now a simple interface for querying images stored in the eMMC,
SD, flash and arbitrary files to find out LOKI metadata that is now
built in to the system FIT image (device tree). This includes
application name and version, LOKI core version, platform and the time
when the image was packaged. This can also be recovered from the
current runtime image (except for the timestamp, since this is in the
FIT top level).

The aim is to give the user more information when updating system
images.

Currently, the metadata is added as an additional dtsi device tree file
included in the build. but the data is a placeholder. This will
eventually replace the metadata previously manually injected into the
filesystem.