Remove old metadata method, fix autoconf configure generation
The old metadata method used separate files for LOKI metadata inserted
into the filesystem. This has been replaced with a version that instead
places the metadata into the device tree, which then appears in the
filesystem automatically while also allowing for the inspection of
images in storage.

While working with autoconf it was noticed that the handling of the
configure file was a little silly since it is a generated file from
configure.ac anyway; this has now been removed.