Add loki-update.sh backup and restore
The --backup and --restore functionality is now available in the LOKI
update utility. This will take an image (from eMMC by default) and back
it up to a single location (actually elsewhere on eMMC), that can be
later restored to eMMC as desired.

The location 'backup' has also been added as a valid target for the
--info flag so that it can be inspected.

Additionally, the --info flag now has an 'alljson' output, which will
output all data about a target in JSON format, to make processing
easier (and to prevent the user from having to make several potentially
costly calls to get data separately).