Fix accidental string casting of config variables, more robust watchdog
For various incoming config variables, the result of kwargs.get() for
numerical values was a string value with a default integer (or other
type) value. This has mean that issues have not appeared until those
variables are used.

This was noticed when the interval time for the environment monitoring
check in the watchdog was a string type, which caused issues later on
where the string value could not be compared. This has now been
corrected for all incoming configuration values, and the watchdog has
been made more robust, casting delays directly to floats to catch issues
earlier. Also, if there are issues checking for kicks of a single
thread, it will carry on to check all threads anyway (this makes
tracking down further issues easier).