Bump core odin-devices to loki-support branch latest
The odin-devices repo already had an override that rather than building
from a tagged version of odin-devices used an out-of-mainline branch
that included an additional device. Each of the application systems was
then using a different override in its own layers to support various
devices, so things have been simplified by using a combined
out-of-mainline branch loki-support.

Ideally the drivers present in this branch will eventually be merged.