Makefile hotfix: constraints rework, moved to design Makefile
The constraints rules are now in the design Makefile, and the
complication of requirements from the hardware/software build are now
more contained within a single Makefile. This is a lot less confusing,
and the simplified rules have allowed the ironing out of issues in
dependency confusion between the project, platform and constraints
rules.