Fix network mount init scripts to support development overrides
Since the latest changes that upgrade the init scripts to allow for
multiple instances of odin-control to run, the network mount has not
been compatible since it was only designed with one instance in mind.

With some modifications, this restores compatibility so that the network
overrides work with systems using multiple instances. Each instance now
has a system configuration file (e.g. loki-update-config.conf) which can
be overidden by specifying it as LOKI_CONFIG_FILENAME in the layout.conf
file. This file only will be overridden. Otherwise, config options in
the existing network mounts have been left as-is, although you do have
to correct the filename.

Every instance will also use the default config unless specified. You
can override settings for all instances by overriding this file, which
is config-default.conf.

In later commits, I intend to allow the user to provide a list of
overridden config names.