Switched init script system to systemd unit files
Originally these scripts were SysV init files. As a quick fix, they have
all been wrapped in systemd unit files. This is not the most robust way
to do this, but will get the system running. In future the SysV parts of
the scripts should be removed, and they should be converted to properly
written systemd units.

The config has also been updated to match the new mountpoint names. This
currently breaks the loki-update instance.